requirements.txt
streamlit==1.31.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.0
lxml==5.3.0
pandas==2.2.2
numpy==1.26.4
sqlite3  # built-ind
